requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.1
pymongo>=4.6.0
gTTS>=2.5.1
azure-cognitiveservices-speech>=1.37.0
google-generativeai>=0.8.5
